+++ api/src/lib/User/index.ts
@@ -1,5 +1,5 @@
-import User from './user';
-import type { UserState, Vector } from './user';
+import User from './user-controller';
+import type { UserState, Vector } from './user-controller';
 
 export type { UserState, Vector };
 export { User };
